server/requirements.txt
aiohttp==1.2.0
peewee==2.8.5
python-socketio==1.7.1
